feat: convert module file to mjs
